encapsulating handleDown & fixing redux auth
