+++ Clients/src/application/tools/fileDownload.ts
@@ -1,0 +1,22 @@
+export const handleDownload = async (fileId: string, fileName: string) => {
+  const token = localStorage.getItem("token");
+  try {
+    const res = await fetch(`http://localhost:3000/files/${fileId}`, {
+      method: "GET",
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+    const blob = await res.blob();
+    const url = window.URL.createObjectURL(blob);
+    const a = document.createElement("a");
+    a.href = url;
+    a.download = fileName;
+    document.body.appendChild(a);
+    a.click();
+    a.remove();
+    URL.revokeObjectURL(url);
+  } catch (error) {
+    console.error("Error downloading file:", error);
+  }
+};
