+++ Clients/src/presentation/components/Table/FilesBasicTable/FileBasicTable.tsx
@@ -12,9 +12,10 @@ import {
 } from "@mui/material";
 import TablePaginationActions from "../../TablePagination";
 import singleTheme from "../../../themes/v1SingleTheme";
-import { useState, useEffect, useCallback} from "react";
+import { useState, useEffect, useCallback } from "react";
 import { FileData } from "../../../../domain/File";
 import IconButton from "../../IconButton";
+import { handleDownload } from "../../../../application/tools/fileDownload";
 
 const DEFAULT_ROWS_PER_PAGE = 5;
 
@@ -78,29 +79,6 @@ const FileBasicTable: React.FC<FileBasicTableProps> = ({
     page * rowsPerPage + rowsPerPage
   );
 
-  //download file
-  const handleDownload = async (fileId: string, fileName:string) => {
-    const token = localStorage.getItem("token");
-    try{
-      const res = await fetch(`http://localhost:3000/files/${fileId}`, {
-        method: "GET",
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
-      const blob = await res.blob();
-      const url = window.URL.createObjectURL(blob);
-      const a = document.createElement("a");
-      a.href = url;
-      a.download = fileName;
-      document.body.appendChild(a);
-      a.click();
-      a.remove();
-      URL.revokeObjectURL(url);
-    } catch (error) {
-      console.error("Error downloading file:", error);
-    }
-  };
   return (
     <>
       <TableContainer id={table}>
@@ -140,16 +118,16 @@ const FileBasicTable: React.FC<FileBasicTableProps> = ({
                 <TableCell>{row.uploadDate}</TableCell>
                 <TableCell>{row.uploader}</TableCell>
                 <TableCell>
-                 <IconButton 
-                 id={Number(row.id)}
-                 type="report"
-                 onEdit={()=> handleDownload(row.id, row.fileName)}
-                 onDelete={() =>{}}
-                 warningTitle="Are you sure you want to download this file?"
-                 warningMessage="This action will download the file to your local machine."
-                 onMouseEvent={()=>{}}
-                  hideRemove={true}
-                 />
+                  <IconButton
+                    id={Number(row.id)}
+                    type="report"
+                    onEdit={() => handleDownload(row.id, row.fileName)}
+                    onDelete={() => {}}
+                    warningTitle="Are you sure you want to download this file?"
+                    warningMessage="This action will download the file to your local machine."
+                    onMouseEvent={() => {}}
+                    hideRemove={true}
+                  />
                 </TableCell>
               </TableRow>
             ))}
